Add files via upload
Added a Folder which Contains a JN File That Explains the Concept of Transfer Learning ( I Used the VGG16 Model) To Classify Flowers of 5 Different Kinds Flowers.
Since the Dataset is Small and also different From the Original Training Dataset, I used the Fine-tuning Approach to fine-tune the last layers of the Model (I,e all FC Layers). The Conv and the Pooling Layers Weights Were not Trained and Freezed During the Training.