Add files via upload
This JN explains how to scrap data from the websites using beautiful-soup which is one of the web scraping frameworks such as selenium etc. This process is the first step in the Text Processing which is basically capturing the data and cleaning(removing messy and unwanted data , noisy data , HTML Tags  etc).